\name{DataCluster1}
\alias{DataCluster1}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataCluster1)
}
\arguments{
  \item{DataCluster1}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataCluster1)
}
