\name{DataCluster2}
\alias{DataCluster2}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataCluster2)
}
\arguments{
  \item{DataCluster2}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataCluster2)
}
