\name{DataCluster3}
\alias{DataCluster3}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataCluster3)
}
\arguments{
  \item{DataCluster3}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataCluster3)
}
