\name{DataCluster4}
\alias{DataCluster4}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataCluster4)
}
\arguments{
  \item{DataCluster4}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataCluster4)
}
