\name{DataCompareStr1}
\alias{DataCompareStr1}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataCompareStr1)
}
\arguments{
  \item{DataCompareStr1}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataCompareStr1)
}
