\name{DataCompareStr2}
\alias{DataCompareStr2}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataCompareStr2)
}
\arguments{
  \item{DataCompareStr2}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataCompareStr2)
}
