\name{DataRNAstr}
\alias{DataRNAstr}
\title{A RNA secondary structure file containing structure information}
\usage{
data(DataRNAstr)
}
\arguments{
  \item{DataRNAstr}{
A RNA secondary structure file containing structure information
}
}
\description{
A RNA secondary structure file containing structure information
}

\author{
Zheng Hewei
}
\examples{
##########
data(DataRNAstr)
}
